hackothon/frontend/src/app/predict/brain/requirements.txt
streamlit>=1.38.0
torch>=2.2.0
torchvision>=0.17.0
pillow>=10.0.0
opencv-python>=4.9.0.0
numpy>=1.26.0
pandas>=2.2.0
matplotlib>=3.8.0
plotly>=5.20.0
transformers>=4.40.0
scikit-learn>=1.4.0
lime>=0.2.0.1
scikit-image>=0.23.0
xgboost>=2.0.0
joblib>=1.3.0
scikit-image>=0.23.0
opencv-python-headless
xgboost>=2.0.0
shap>=0.46.0
scikit-learn>=1.4.0
scikit-image>=0.23.0
